python/requirements.txt
# Python Requirements for CV Analysis and Performance Prediction
# تثبيت المكتبات باستخدام: pip install -r requirements.txt

# استخراج النص من ملفات PDF
pdfplumber==0.10.4

# استخراج النص من ملفات DOCX
python-docx==1.1.0

# Machine Learning للمحلل الذكي
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.24.3
joblib==1.3.2

# معالجة النصوص (اختياري)
# nltk==3.8.1

# المكتبات المدمجة (لا تحتاج تثبيت):
# json
# sys
# re
# os
# pathlib

